fix: add check options

+++ src/server/api/databaseDelete.ts
@@ -1,10 +1,18 @@
 import type { Context } from 'hono'
 
 import { connectClient } from '@/server/db'
+import { checkOptions } from '@/utils/validations/serverChecks'
 
 export default async function databaseDelete(c: Context) {
   const { database } = await c.req.json<{ database: string }>()
   await connectClient()
+  const { error: optionError } = checkOptions({
+    readOnly: false,
+    noDelete: false
+  })
+  if (optionError) {
+    return c.json({ error: optionError }, 403)
+  }
   await globalThis.mongo.mongoClient.db(database).dropDatabase()
     .catch((error) => {
       console.debug(error)
